Moved populate_edges into TypedGraphBuilder
